radius=int(input()) area=3.14*radius**2 print(area)

+++ Area_of_a_circle.py
@@ -1,4 +1,3 @@
-radius=float(input())
+radius=int(input())
 area=3.14*radius**2
-
 print(area)
